backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
chatgpt==2.2212.0
click==8.3.1
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
nltk==3.9.2
openai==2.15.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyphen==0.17.2
python-dotenv==1.2.1
regex==2026.1.15
rich==14.2.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
textstat==0.7.12
tls-client==1.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
